django_mooc/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
Django==3.0.1
django-filter==2.2.0
django-guardian==2.2.0
django-simpleui==3.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itypes==1.1.0
Jinja2==2.11.1
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
parsel==1.5.2
Pillow==7.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==2.0.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
uritemplate==3.0.1
urllib3==1.25.8
w3lib==1.21.0
zope.interface==4.7.1
